Moved Java file. Finished Parser. Recreated Lexical Analyzer using C#
Had issues with java Lexical Analyzer, so had to create a new project so
that it would work again. Deleted old version of it. Was having trouble
with makefile to turn in as it was supposed to compile 2 different
languages. Since was having issues, rewrote Java Lexical Analyzer into
C#. Works as intended, but might have issues when it comes to error
reporting. Might look into later to fix.